# src/main/java/com/probestack/forgestudio/design/api/InvoicesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CreateInvoiceRequest;
import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.Invoice;
import com.probestack.forgestudio.design.model.InvoiceListResponse;
import com.probestack.forgestudio.design.model.UpdateInvoiceRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-14T08:34:39.853488103Z[GMT]")@Validated
@Tag(name = "Invoices", description = "Operations on customer invoices")
public interface InvoicesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /invoices : Create a new invoice
     * Creates a draft invoice from a set of line items. The &#x60;invoiceNumber&#x60; is generated by the server and is unique.
     *
     * @param createInvoiceRequest  (required)
     * @return Invoice created successfully (status code 201)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or An invoice already exists for this billing period (status code 409)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "createInvoice",
        summary = "Create a new invoice",
        description = "Creates a draft invoice from a set of line items. The `invoiceNumber` is generated by the server and is unique.",
        tags = { "Invoices" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Invoice created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Invoice.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "An invoice already exists for this billing period", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/invoices",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Invoice> createInvoice(
        @Parameter(name = "CreateInvoiceRequest", description = "", required = true) @Valid @RequestBody CreateInvoiceRequest createInvoiceRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"dueDate\" : \"2026-08-31T00:00:00.000+00:00\", \"lineItems\" : [ { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" }, { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" } ], \"totalAmount\" : 1416.0, \"createdAt\" : \"2026-08-01T09:00:00Z\", \"amountPaid\" : 0, \"subtotal\" : 1200.0, \"invoiceNumber\" : \"INV-2026-00317\", \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"i1n2v3c4\", \"issuedAt\" : \"2026-08-01T00:00:00.000+00:00\", \"taxAmount\" : 216.0, \"status\" : \"ISSUED\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /invoices/{invoiceId} : Get an invoice by ID
     * Retrieves the full details of a single invoice including line items and tax breakdown.
     *
     * @param invoiceId Unique invoice identifier (UUID) (required)
     * @return Invoice found (status code 200)
     *         or Invoice not found (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "getInvoiceById",
        summary = "Get an invoice by ID",
        description = "Retrieves the full details of a single invoice including line items and tax breakdown.",
        tags = { "Invoices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Invoice found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Invoice.class))
            }),
            @ApiResponse(responseCode = "404", description = "Invoice not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/invoices/{invoiceId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Invoice> getInvoiceById(
        @Parameter(name = "invoiceId", description = "Unique invoice identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("invoiceId") String invoiceId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"dueDate\" : \"2026-08-31T00:00:00.000+00:00\", \"lineItems\" : [ { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" }, { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" } ], \"totalAmount\" : 1416.0, \"createdAt\" : \"2026-08-01T09:00:00Z\", \"amountPaid\" : 0, \"subtotal\" : 1200.0, \"invoiceNumber\" : \"INV-2026-00317\", \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"i1n2v3c4\", \"issuedAt\" : \"2026-08-01T00:00:00.000+00:00\", \"taxAmount\" : 216.0, \"status\" : \"ISSUED\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /invoices : List all invoices
     * Returns a paginated list of invoices, optionally filtered by status or customer. Supports standard pagination parameters (page, size).
     *
     * @param customerId Filter by the billed customer (optional)
     * @param status Filter by invoice status (optional)
     * @param page Page number (starts at 1) (optional, default to 1)
     * @param size Number of items per page (optional, default to 20)
     * @return Successful response (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "listInvoices",
        summary = "List all invoices",
        description = "Returns a paginated list of invoices, optionally filtered by status or customer. Supports standard pagination parameters (page, size).",
        tags = { "Invoices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InvoiceListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/invoices",
        produces = { "application/json" }
    )
    
    default ResponseEntity<InvoiceListResponse> listInvoices(
        @Parameter(name = "customerId", description = "Filter by the billed customer", in = ParameterIn.QUERY) @Valid @RequestParam(value = "customerId", required = false) String customerId,
        @Parameter(name = "status", description = "Filter by invoice status", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) String status,
        @Min(1) @Parameter(name = "page", description = "Page number (starts at 1)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "Number of items per page", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"size\" : 20, \"totalPages\" : 16, \"page\" : 1, \"content\" : [ { \"dueDate\" : \"2026-08-31T00:00:00.000+00:00\", \"lineItems\" : [ { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" }, { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" } ], \"totalAmount\" : 1416.0, \"createdAt\" : \"2026-08-01T09:00:00Z\", \"amountPaid\" : 0, \"subtotal\" : 1200.0, \"invoiceNumber\" : \"INV-2026-00317\", \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"i1n2v3c4\", \"issuedAt\" : \"2026-08-01T00:00:00.000+00:00\", \"taxAmount\" : 216.0, \"status\" : \"ISSUED\" }, { \"dueDate\" : \"2026-08-31T00:00:00.000+00:00\", \"lineItems\" : [ { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" }, { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" } ], \"totalAmount\" : 1416.0, \"createdAt\" : \"2026-08-01T09:00:00Z\", \"amountPaid\" : 0, \"subtotal\" : 1200.0, \"invoiceNumber\" : \"INV-2026-00317\", \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"i1n2v3c4\", \"issuedAt\" : \"2026-08-01T00:00:00.000+00:00\", \"taxAmount\" : 216.0, \"status\" : \"ISSUED\" } ], \"totalElements\" : 317 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * PUT /invoices/{invoiceId} : Update an existing invoice
     * Replaces the editable fields of a draft invoice. Issued invoices can only have their due date changed.
     *
     * @param invoiceId Unique invoice identifier (UUID) (required)
     * @param updateInvoiceRequest  (required)
     * @return Invoice updated successfully (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "updateInvoice",
        summary = "Update an existing invoice",
        description = "Replaces the editable fields of a draft invoice. Issued invoices can only have their due date changed.",
        tags = { "Invoices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Invoice updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Invoice.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/invoices/{invoiceId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Invoice> updateInvoice(
        @Parameter(name = "invoiceId", description = "Unique invoice identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("invoiceId") String invoiceId,
        @Parameter(name = "UpdateInvoiceRequest", description = "", required = true) @Valid @RequestBody UpdateInvoiceRequest updateInvoiceRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"dueDate\" : \"2026-08-31T00:00:00.000+00:00\", \"lineItems\" : [ { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" }, { \"unitPrice\" : 100.0, \"quantity\" : 12, \"taxRatePercent\" : 18.0, \"description\" : \"Platform subscription - August 2026\" } ], \"totalAmount\" : 1416.0, \"createdAt\" : \"2026-08-01T09:00:00Z\", \"amountPaid\" : 0, \"subtotal\" : 1200.0, \"invoiceNumber\" : \"INV-2026-00317\", \"customerId\" : \"cust-88213\", \"currency\" : \"USD\", \"id\" : \"i1n2v3c4\", \"issuedAt\" : \"2026-08-01T00:00:00.000+00:00\", \"taxAmount\" : 216.0, \"status\" : \"ISSUED\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * DELETE /invoices/{invoiceId} : Void an invoice
     * Voids the invoice so it no longer counts towards receivables. Paid invoices cannot be voided.
     *
     * @param invoiceId Unique invoice identifier (UUID) (required)
     * @return Invoice voided successfully (no content returned) (status code 204)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "voidInvoice",
        summary = "Void an invoice",
        description = "Voids the invoice so it no longer counts towards receivables. Paid invoices cannot be voided.",
        tags = { "Invoices" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Invoice voided successfully (no content returned)"),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/invoices/{invoiceId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> voidInvoice(
        @Parameter(name = "invoiceId", description = "Unique invoice identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("invoiceId") String invoiceId
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }

}